+++ packages/client/src/constants/routes.ts
@@ -5,5 +5,4 @@ export enum RoutePath {
   Game = 'game',
   Leaderboards = 'leaderboards',
   Forum = 'forum',
-  Profile = 'profile',
 }
